Modificacion al comp de Order y cambio a auth con Cookies

+++ src/app/Components/order-get/order-get.component.ts
@@ -39,9 +39,11 @@ export class OrderGetComponent implements OnInit {
   }
 
   deleteOrder(id){
-    this.orderService.deleteDinningRoom(id).subscribe(()=>{
-      this.fetchOrders();
-    })
+    if(confirm('Are you sure to delete this record?')){
+      this.orderService.deleteDinningRoom(id).subscribe(()=>{
+        this.fetchOrders();
+      })
+    }
   }
 
 }
